requirements.txt
requests
python-dotenv
colorama
anthropic
emoji
